+++ modules/Students/Search.inc.php
@@ -334,6 +334,20 @@ else
 
 		$extra['options'] = issetVal( $extra['options'], [] );
 
+		if ( empty( $extra['columns_before'] )
+			&& ( empty( $extra['link']['remove'] ) || ! AllowEdit() ) )
+		{
+			// @since 12.6 CSS Fix first column (list, table)
+			if ( ! empty( $extra['options']['class'] ) )
+			{
+				$extra['options']['class'] .= ' fix-first-column';
+			}
+			else
+			{
+				$extra['options']['class'] = 'fix-first-column';
+			}
+		}
+
 		if ( ! empty( $_REQUEST['address_group'] ) )
 		{
 			ListOutput(